Make setting monitors_locked an unsigned long (#1196)
The lock counter was always used as a non-negative integer and should
also be such on the type level. In the original settings code, all
integers where signed, but now it is finally updated.

Also it is not necessary anymore to call lock_number_changed() manually,
because this happens automatically through signals.

Both changes are tested by the new test cases for lock/unlock.

+++ src/settings.h
@@ -60,7 +60,7 @@ public:
     Attribute_<bool>          hide_covered_windows = {"hide_covered_windows", false};
     Attribute_<bool>          smart_frame_surroundings = {"smart_frame_surroundings", false};
     Attribute_<bool>          smart_window_surroundings = {"smart_window_surroundings", false};
-    Attribute_<int>           monitors_locked = {"monitors_locked", 0};
+    Attribute_<unsigned long> monitors_locked = {"monitors_locked", 0};
     Attribute_<bool>          auto_detect_monitors = {"auto_detect_monitors", false};
     Attribute_<bool>          auto_detect_panels = {"auto_detect_panels", true};
     Attribute_<int>           pseudotile_center_threshold = {"pseudotile_center_threshold", 10};
